+++ packages/providers/lib/fallback-provider.d.ts
@@ -1,5 +1,5 @@
-import { Network } from "@ethersproject/networks";
 import { Provider } from "@ethersproject/abstract-provider";
+import { Network } from "@ethersproject/networks";
 import { BaseProvider } from "./base-provider";
 export interface FallbackProviderConfig {
     provider: Provider;